vector_db/requirements.txt
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
